Adds PHPUnit 11 support
