+++ .github/workflows/tests.yml
@@ -17,13 +17,14 @@ jobs:
     strategy:
       fail-fast: true
       matrix:
-        php: [8.1, 8.2, 8.3]
+        php: [8.2, 8.3]
+        phpunit: [10, 11]
         laravel: [10, 11]
         exclude:
-          - php: 8.1
-            laravel: 11
+          - phpunit: 11
+            laravel: 10
 
-    name: PHP ${{ matrix.php }} - Laravel ${{ matrix.laravel }}
+    name: PHP ${{ matrix.php }} - PHPUnit ${{ matrix.phpunit }} - Laravel ${{ matrix.laravel }}
 
     steps:
       - name: Checkout code
@@ -40,6 +41,7 @@ jobs:
 
       - name: Install dependencies
         run: |
+          composer require phpunit/phpunit:^${{ matrix.phpunit }} --no-update
           composer require "laravel/framework=^${{ matrix.laravel }}" --no-update
           composer update --prefer-dist --no-interaction --no-progress
