+++ composer.json
@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "ext-dom": "*",
         "illuminate/contracts": "^10.0|^11.0",
         "illuminate/database": "^10.0|^11.0",
@@ -18,7 +18,7 @@
         "illuminate/support": "^10.0|^11.0",
         "illuminate/testing": "^10.0|^11.0",
         "mockery/mockery": "^1.0",
-        "phpunit/phpunit": "^10.0.7",
+        "phpunit/phpunit": "^10.4|^11.0",
         "symfony/console": "^6.2|^7.0",
         "symfony/css-selector": "^6.2|^7.0",
         "symfony/dom-crawler": "^6.2|^7.0",
