+++ src/Constraints/Concerns/PageConstraint.php
@@ -1,0 +1,123 @@
+<?php
+
+namespace Laravel\BrowserKitTesting\Constraints\Concerns;
+
+use PHPUnit\Framework\ExpectationFailedException;
+use SebastianBergmann\Comparator\ComparisonFailure;
+use Symfony\Component\DomCrawler\Crawler;
+
+trait PageConstraint
+{
+    /**
+     * Make sure we obtain the HTML from the crawler or the response.
+     *
+     * @param  \Symfony\Component\DomCrawler\Crawler|string  $crawler
+     * @return string
+     */
+    protected function html($crawler)
+    {
+        return is_object($crawler) ? $crawler->html() : $crawler;
+    }
+
+    /**
+     * Make sure we obtain the HTML from the crawler or the response.
+     *
+     * @param  \Symfony\Component\DomCrawler\Crawler|string  $crawler
+     * @return string
+     */
+    protected function text($crawler)
+    {
+        return is_object($crawler) ? $crawler->text() : strip_tags($crawler);
+    }
+
+    /**
+     * Create a crawler instance if the given value is not already a Crawler.
+     *
+     * @param  \Symfony\Component\DomCrawler\Crawler|string  $crawler
+     * @return \Symfony\Component\DomCrawler\Crawler
+     */
+    protected function crawler($crawler)
+    {
+        return is_object($crawler) ? $crawler : new Crawler($crawler);
+    }
+
+    /**
+     * Get the escaped text pattern for the constraint.
+     *
+     * @param  string  $text
+     * @return string
+     */
+    protected function getEscapedPattern($text)
+    {
+        $rawPattern = preg_quote($text, '/');
+
+        $escapedPattern = preg_quote(e($text), '/');
+
+        return $rawPattern == $escapedPattern
+            ? $rawPattern : "({$rawPattern}|{$escapedPattern})";
+    }
+
+    /**
+     * Throw an exception for the given comparison and test description.
+     *
+     * @param  mixed  $other
+     * @param  string  $description
+     * @param  \SebastianBergmann\Comparator\ComparisonFailure|null  $comparisonFailure
+     * @return void
+     *
+     * @throws \PHPUnit\Framework\ExpectationFailedException
+     */
+    protected function fail(mixed $other, string $description, ?ComparisonFailure $comparisonFailure = null): never
+    {
+        $html = $this->html($other);
+
+        $failureDescription = sprintf(
+            "%s\n\n\nFailed asserting that %s",
+            $html, $this->getFailureDescription()
+        );
+
+        if (! empty($description)) {
+            $failureDescription .= ": {$description}";
+        }
+
+        if (trim($html) != '') {
+            $failureDescription .= '. Please check the content above.';
+        } else {
+            $failureDescription .= '. The response is empty.';
+        }
+
+        throw new ExpectationFailedException($failureDescription, $comparisonFailure);
+    }
+
+    /**
+     * Get the description of the failure.
+     *
+     * @return string
+     */
+    protected function getFailureDescription()
+    {
+        return 'the page contains '.$this->toString();
+    }
+
+    /**
+     * Returns the reversed description of the failure.
+     *
+     * @return string
+     */
+    protected function getReverseFailureDescription()
+    {
+        return 'the page does not contain '.$this->toString();
+    }
+
+    /**
+     * Get a string representation of the object.
+     *
+     * Placeholder method to avoid forcing definition of this method.
+     *
+     * @return string
+     */
+    public function toString(): string
+    {
+        return '';
+    }
+}
