+++ src/Constraints/HasText.php
@@ -2,46 +2,16 @@
 
 namespace Laravel\BrowserKitTesting\Constraints;
 
-class HasText extends PageConstraint
-{
-    /**
-     * The expected text.
-     *
-     * @var string
-     */
-    protected $text;
+use PHPUnit\Runner\Version;
 
-    /**
-     * Create a new constraint instance.
-     *
-     * @param  string  $text
-     * @return void
-     */
-    public function __construct($text)
+if (str_starts_with(Version::series(), '10')) {
+    class HasText extends PageConstraint
     {
-        $this->text = $text;
+        use Concerns\HasText;
     }
-
-    /**
-     * Check if the plain text is found in the given crawler.
-     *
-     * @param  \Symfony\Component\DomCrawler\Crawler|string  $crawler
-     * @return bool
-     */
-    protected function matches($crawler): bool
+} else {
+    readonly class HasText extends PageConstraint
     {
-        $pattern = $this->getEscapedPattern($this->text);
-
-        return preg_match("/{$pattern}/i", $this->text($crawler));
-    }
-
-    /**
-     * Returns a string representation of the object.
-     *
-     * @return string
-     */
-    public function toString(): string
-    {
-        return "the text [{$this->text}]";
+        use Concerns\HasText;
     }
 }
