+++ src/Constraints/HasValue.php
@@ -2,73 +2,16 @@
 
 namespace Laravel\BrowserKitTesting\Constraints;
 
-use Symfony\Component\DomCrawler\Crawler;
+use PHPUnit\Runner\Version;
 
-class HasValue extends FormFieldConstraint
-{
-    /**
-     * Get the valid elements.
-     *
-     * @return string
-     */
-    protected function validElements()
+if (str_starts_with(Version::series(), '10')) {
+    class HasValue extends FormFieldConstraint
     {
-        return 'input,textarea';
+        use Concerns\HasValue;
     }
-
-    /**
-     * Check if the input contains the expected value.
-     *
-     * @param  \Symfony\Component\DomCrawler\Crawler|string  $crawler
-     * @return bool
-     */
-    public function matches($crawler): bool
+} else {
+    readonly class HasValue extends FormFieldConstraint
     {
-        $crawler = $this->crawler($crawler);
-
-        return $this->getInputOrTextAreaValue($crawler) == $this->value;
-    }
-
-    /**
-     * Get the value of an input or textarea.
-     *
-     * @param  \Symfony\Component\DomCrawler\Crawler  $crawler
-     * @return string
-     *
-     * @throws \PHPUnit\Framework\ExpectationFailedException
-     */
-    public function getInputOrTextAreaValue(Crawler $crawler)
-    {
-        $field = $this->field($crawler);
-
-        return $field->nodeName() == 'input'
-            ? $field->attr('value')
-            : $field->text();
-    }
-
-    /**
-     * Return the description of the failure.
-     *
-     * @return string
-     */
-    protected function getFailureDescription()
-    {
-        return sprintf(
-            'the field [%s] contains the expected value [%s]',
-            $this->selector, $this->value
-        );
-    }
-
-    /**
-     * Returns the reversed description of the failure.
-     *
-     * @return string
-     */
-    protected function getReverseFailureDescription()
-    {
-        return sprintf(
-            'the field [%s] does not contain the expected value [%s]',
-            $this->selector, $this->value
-        );
+        use Concerns\HasValue;
     }
 }
