+++ src/Constraints/ReversePageConstraint.php
@@ -2,56 +2,16 @@
 
 namespace Laravel\BrowserKitTesting\Constraints;
 
-class ReversePageConstraint extends PageConstraint
-{
-    /**
-     * The page constraint instance.
-     *
-     * @var \Laravel\BrowserKitTesting\Constraints\PageConstraint
-     */
-    protected $pageConstraint;
+use PHPUnit\Runner\Version;
 
-    /**
-     * Create a new reverse page constraint instance.
-     *
-     * @param  \Laravel\BrowserKitTesting\Constraints\PageConstraint  $pageConstraint
-     * @return void
-     */
-    public function __construct(PageConstraint $pageConstraint)
+if (str_starts_with(Version::series(), '10')) {
+    class ReversePageConstraint extends PageConstraint
     {
-        $this->pageConstraint = $pageConstraint;
+        use Concerns\ReversePageConstraint;
     }
-
-    /**
-     * Reverse the original page constraint result.
-     *
-     * @param  \Symfony\Component\DomCrawler\Crawler  $crawler
-     * @return bool
-     */
-    public function matches($crawler): bool
+} else {
+    readonly class ReversePageConstraint extends PageConstraint
     {
-        return ! $this->pageConstraint->matches($crawler);
-    }
-
-    /**
-     * Get the description of the failure.
-     *
-     * This method will attempt to negate the original description.
-     *
-     * @return string
-     */
-    protected function getFailureDescription()
-    {
-        return $this->pageConstraint->getReverseFailureDescription();
-    }
-
-    /**
-     * Get a string representation of the object.
-     *
-     * @return string
-     */
-    public function toString(): string
-    {
-        return $this->pageConstraint->toString();
+        use Concerns\ReversePageConstraint;
     }
 }
